+++ pages/lex/units/+Page.ts
@@ -4,21 +4,9 @@ import { ContentPage } from "~/layouts";
 import { useState, useEffect } from "react";
 import { fetchAPIData } from "~/_utils";
 import { usePageContext } from "vike-react/usePageContext";
-import { ClientOnly } from "vike-react/ClientOnly";
-import { FlexRow } from "@macrostrat/ui-components";
 import { PostgRESTInfiniteScrollView } from "@macrostrat/ui-components";
 import { postgrestPrefix } from "@macrostrat-web/settings";
-
-function LithologyTag(props) {
-  return h(
-    ClientOnly,
-    {
-      load: () => import("~/components/lex/lithology-tag.client").then((d) => d.LithologyTagInner),
-      deps: [props.data, props.href],
-    },
-    (component) => h(component, props)
-  );
-}
+import { LithologyTag, FlexRow } from "~/components/lex/tag";
 
 export function Page() {
   const url = usePageContext().urlOriginal.split("?")[1];
@@ -36,27 +24,28 @@ export function Page() {
   const name = params.name;
 
   useEffect(() => {
-    fetchAPIData("/fossils", {
+    fetchAPIData("/units", {
       [idType]: id
     }).then((res) => setData(res));
   }, []);
 
+
   return h(ContentPage, [
     h(Header, { name, color, idType, id }),
     h('div.units', [
-      data?.map((d) => h(FossilItem, { key: d.id, data: d })),
+      data?.map((d) => h(UnitItem, { key: d.id, data: d })),
     ]),
     h(Footer)
   ]);
 }
 
-function FossilItem({ data }) {
-  const { cltn_name, cltn_id } = data;
+function UnitItem({ data }) {
+  const { unit_id, col_id, unit_name } = data;
 
   return h(LinkCard, {
-    href: 'https://paleobiodb.org/classic/displayCollResults?collection_no=col:' + cltn_id,
-    className: "fossil-item",
-    title: cltn_name,
+    href: `/columns/${col_id}#unit=${unit_id}`,
+    className: "unit-item",
+    title: unit_name,
   });
 }
 
@@ -67,13 +56,12 @@ function Header({ name, color, idType, id }) {
     'econ_id': "economics",
     'environ_id': "environments",
     'strat_name_id': "strat-names",
-    'strat_name_concept_id': "strat-concepts",
   }
 
   return h(StickyHeader, { className: "header" }, [
     h(PageBreadcrumbs, {
       title: h(FlexRow, { gap: ".5em", alignItems: "center" }, [
-        h('p.title', 'Fossils for '),
+        h('p.title', 'Units for '),
         h(LithologyTag, { data: { name, color }, href: `/lex/${map[idType]}/${id}` }),
       ]),
     })
@@ -97,24 +85,23 @@ function getUrlParams(urlString) {
 
 function Base() {
   return h(ContentPage, { className: 'page' }, [
-    h(StickyHeader, { className: "header" }, h(PageBreadcrumbs, { title: "Fossils" })),
+    h(StickyHeader, { className: "header" }, h(PageBreadcrumbs, { title: "Units" })),
     h(PostgRESTInfiniteScrollView, {
-      route: postgrestPrefix + '/fossils',
-      id_key: 'collection_no',
+      route: postgrestPrefix + '/units',
+      id_key: 'id',
       limit: 20,
-      itemComponent: BaseFossilItem,
+      itemComponent: BaseUnitItem,
       filterable: true,
-      searchColumns: [{ value: 'name', label: 'Fossil Name' }],
+      searchColumns: [{ value: "strat_name", label: "Name" }],
     }),
   ]);
 }
 
-function BaseFossilItem({ data }) {
-  const { name, collection_no } = data;
+function BaseUnitItem({ data }) {
+  const { id, col_id, strat_name } = data;
 
   return h(LinkCard, {
-    href: 'https://paleobiodb.org/classic/displayCollResults?collection_no=col:' + collection_no,
-    className: "fossil-item",
-    title: name,
-  });
+    href: `/columns/${col_id}#unit=${id}`,
+    title: strat_name,
+  })
 }
