+++ src/components/lex/load.client.ts
@@ -1,5 +1,6 @@
 import h from "@macrostrat/hyper";
 import { LithologyTag } from "@macrostrat/data-components";
+import { FlexRow } from "@macrostrat/ui-components";
 
 export function LithologyTagInner({ data, href }) {
   return h(
@@ -9,4 +10,8 @@ export function LithologyTagInner({ data, href }) {
       data
     },
   );
+}
+
+export function FlexRowInner(props) {
+  return h(FlexRow, props);
 }
